+++ .husky/pre-commit
@@ -1,4 +1,4 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-./node_modules/.bin/npm-run-all preCommit test linters postinstall
+./node_modules/.bin/npm-run-all preCommit
