+++ apiExamples/basic.html
@@ -1,23 +1,12 @@
-<auro-slideshow loop pagination variant="slider">
+<auro-slideshow loop pagination>
   <span slot="header">Slider Header</span>
   <span slot="subheader">Subheader</span>
   <img src="https://picsum.photos/200?random=0" style="height: 300px; width: 600px;" alt="Random insert 0">
-  <div style="height: 300px; width: 200px; background-color: bisque;"></div>
+  <div style="height: 300px; width: 500px; background-color: bisque;">Slide 2</div>
   <img src="https://picsum.photos/200?random=1" style="height: 300px; width: 200px;" alt="Random insert 1">
   <img src="https://picsum.photos/200?random=2" style="height: 300px; width: 300px;" alt="Random insert 2">
   <img src="https://picsum.photos/200?random=3" style="height: 300px; width: 500px;" alt="Random insert 3">
   <img src="https://picsum.photos/200?random=4" style="height: 300px; width: 400px;" alt="Random insert 4">
   <img src="https://picsum.photos/200?random=5" style="height: 300px; width: 300px;" alt="Random insert 5">
 </auro-slideshow>
-<hr>
-<auro-slideshow loop pagination variant="slideshow">
-  <span slot="header">Slideshow Header</span>
-  <span slot="subheader">Subheader</span>
-  <img src="https://picsum.photos/200?random=0" style="height: 300px; width: 600px;" alt="Random insert 0">
-  <div style="height: 300px; width: 200px; background-color: bisque;"></div>
-  <img src="https://picsum.photos/200?random=1" style="height: 300px; width: 200px;" alt="Random insert 1">
-  <img src="https://picsum.photos/200?random=2" style="height: 300px; width: 300px;" alt="Random insert 2">
-  <img src="https://picsum.photos/200?random=3" style="height: 300px; width: 500px;" alt="Random insert 3">
-  <img src="https://picsum.photos/200?random=4" style="height: 300px; width: 400px;" alt="Random insert 4">
-  <img src="https://picsum.photos/200?random=5" style="height: 300px; width: 300px;" alt="Random insert 5">
-</auro-slideshow>
+
