+++ demo/api.js
@@ -1,14 +1,1 @@
-import "../index.js";
-
-// export function initExamples(initCount = 0) {
-//   try {
-
-//   } catch (err) {
-//     if (initCount <= 20) {
-//       // setTimeout handles issue where content is sometimes loaded after the functions get called
-//       setTimeout(() => {
-//         initExamples(initCount + 1);
-//       }, 100);
-//     }
-//   }
-// }
+import "../src/index.js";
