+++ demo/demo-index.js
@@ -1,0 +1,5 @@
+import { AuroSlideshow } from "../src/auro-slideshow";
+
+AuroSlideshow.register();
+AuroSlideshow.register("custom-slideshow");
+
