--- index.js
@@ -1,3 +1,0 @@
-import { AuroSlideshow } from "./src/auro-slideshow.js";
-
-AuroSlideshow.register();
