+++ package.json
@@ -31,6 +31,7 @@
     "swiper": "^11.2.6"
   },
   "devDependencies": {
+    "@aurodesignsystem/auro-cli": "^2.2.0",
     "@aurodesignsystem/auro-config": "^1.0.1",
     "@aurodesignsystem/design-tokens": "^5.3.1",
     "@aurodesignsystem/webcorestylesheets": "^6.5.0",
@@ -87,7 +88,10 @@
     "last 2 Safari major versions"
   ],
   "nodemonConfig": {
-    "ignore": ["*-css.js", "*.css"]
+    "ignore": [
+      "*-css.js",
+      "*.css"
+    ]
   },
   "release": {
     "branches": [
@@ -129,36 +133,38 @@
   "publishConfig": {
     "access": "public"
   },
-  "keywords": ["alaska airlines", "auro", "design system", "web components"],
+  "keywords": [
+    "alaska airlines",
+    "auro",
+    "design system",
+    "web components"
+  ],
   "scripts": {
-    "build": "npm-run-all build:version build:sass sass:render bundler types",
-    "build:test": "npm-run-all test linters",
-    "build:release": "npm-run-all build build:test build:api build:docs bundler postinstall",
-    "build:ci": "npm-run-all sweep build:release",
-    "build:api": "wca analyze 'scripts/wca/auro-slideshow.js' --outFiles docs/api.md",
-    "build:dev:assets": "npm-run-all build:sass:component postCss:component sass:render build:docs",
-    "build:docs": "node ./node_modules/@aurodesignsystem/auro-library/scripts/build/generateDocs.mjs",
-    "build:sass": "npm-run-all build:sass:component postCss:component sass:render",
-    "build:sass:component": "sass --no-source-map src:src",
-    "build:version": "node scripts/version.mjs",
-    "build:watch": "nodemon -e scss,js,html --watch src --watch apiExamples/**/* --exec npm run build:dev:assets",
-    "bundler": "rollup -c",
-    "bundler:test": "rollup -c -w",
-    "scssLint": "stylelint \"./src/**/*.scss\"",
-    "dev": "concurrently --kill-others \"npm run build:watch\" \"npm run serve\"",
-    "dist:js": "copyfiles -u 1 -V './src/**/*.js' './apiExamples/**/*.js' ./dist",
+    "build": "auro build",
+    "dev": "auro dev --open",
     "lint": "biome check --fix --no-errors-on-unmatched && stylelint \"./src/**/*.scss\"",
-    "linters": "npm-run-all scssLint",
     "preCommit": "node ./node_modules/@aurodesignsystem/auro-library/scripts/build/pre-commit.mjs",
-    "postCss:component": "node ./node_modules/@aurodesignsystem/auro-library/scripts/build/postCss.mjs",
-    "postinstall": "node packageScripts/postinstall.mjs",
-    "sass:render": "sass-render src/*.css -t ./node_modules/@aurodesignsystem/auro-library/scripts/build/staticStyles-template.js",
-    "serve": "web-dev-server --open demo/ --node-resolve --watch",
-    "sweep": "rm -rf ./demo/css ./dist | rm ./src/*.css ./src/*-css.js",
-    "test": "wtr --coverage",
+    "test": "wtr",
     "test:watch": "wtr --watch",
-    "types": "tsc",
-    "prepare": "husky install",
-    "postversion": "npm-run-all build:api build:docs"
-  }
+    "prepare": "husky install"
+  },
+  "exports": {
+    "./readme.md": "./README.md",
+    ".": "./dist/registered.js",
+    "./demo/*.md": "./demo/*.md",
+    "./demo/*.js": "./demo/*.min.js",
+    "./class": {
+      "module": "./dist/index.js",
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist/**/*",
+    "demo/**/*",
+    "CHANGELOG.md",
+    "README.md",
+    "LICENSE",
+    "NOTICE"
+  ]
 }
