+++ src/index.js
@@ -1,0 +1,3 @@
+import { AuroSlideshow } from "./auro-slideshow";
+
+export { AuroSlideshow };
