+++ src/registered.js
@@ -1,0 +1,3 @@
+import { AuroSlideshow } from "./auro-slideshow.js";
+
+AuroSlideshow.register();
